Move get_api_*() to utils
