+++ src/lib/api/utils.server.js
@@ -1,10 +1,38 @@
-import { get_api_auth, get_api_url } from './wp-fetch.server';
+import { env } from '$env/dynamic/private';
 import { get_single_media, get_post_terms, get_posts } from '@kucrut/wp-api-helpers';
 import { process_post_data } from '$lib/utils/post';
 import { process_taxonomy_data } from '$lib/utils/taxonomy';
 import { process_term_data } from '$lib/utils/term';
 
 /** @typedef {(import('$types').Favicon|import('$types').TileImage)[]} Icons */
+
+/**
+ * Get WP API auth
+ *
+ * @return {string} WP API auth string.
+ */
+export function get_api_auth() {
+	const auth = env.WP_API_APP_AUTH || '';
+
+	if ( ! auth ) {
+		return auth;
+	}
+
+	if ( env.WP_API_APP_AUTH_TYPE === 'jwt' ) {
+		return `Bearer ${ auth }`;
+	}
+
+	return `Basic ${ Buffer.from( auth ).toString( 'base64' ) }`;
+}
+
+/**
+ * Get WP API root URL
+ *
+ * @return {string} WP Rest API root URL.
+ */
+export function get_api_url() {
+	return env.WP_API_ROOT_URL;
+}
 
 /**
  * Generate favicons
